feat: add testing libs and created a test for the query 'hello'
